==> 00-data/code/data-cleaning/election-data/00-cleaning_2015_disco.Rmd <==
---
title: "Cleaning 2015 District Council Election"
author: "Jiayi Li"
date: "2022-09-13"
output: html_document
---
Purpose: to clean the election results of the 2015 HK District Council Election at the polling station level.
To-Do: 
- 09/13:
 -- Merge all excel files (done)
 -- Add a constituency column (done)
 -- Geocode the merged file in Python 
 -- Check for bad coordinates 
 -- Reorder columns 
 

# load packages and data

```{r}
# load packages
  source("helper-packages.R")

# load raw election data
  path <- "/Users/jiayili/Dropbox/senior-thesis/00-data/00-raw-data/election-data/2015_district_council_election/individual_stations/"
# get file names
  data_files <- list.files(path, full.names = TRUE)
  
# loop in all file names to load data
  for(i in 1:length(data_files)) {                              
    assign(paste0("hk_2015_disco_", i),                                   
           rio::import(data_files[i]))
  }
  
# load data with candidate names
  path_1 <- "/Users/jiayili/Dropbox/senior-thesis/00-data/00-raw-data/election-data/2015_district_council_election"
  candid_hk_2015_disco <-
    read.csv(glue("{path_1}/hk2015_district_election_candidate_shares.csv")) %>%
    # drop the index column
    select(-X)
```

# reshape the candidate data for matching

## Notes: "*" means uncontested --> dropped

```{r}
# pivot wider
reshaped_hk_2015_disco_candid <-
  candid_hk_2015_disco %>%
  pivot_wider(
    id_cols = `Constituency.Code`,
    names_from = c(`Candidate.Number`),
    values_from = `Name.of.Candidate`,
    names_prefix = "name_candid_"
  ) %>%
  select("district_code" = `Constituency.Code`, starts_with("name_candid_")) %>%
  # drop uncontested candidates
  select(-`name_candid_*`)
```

# stack counting results at the ps level and clean

```{r}
# bind rows 
  hk_2015_dsc_raw <-
    # merge all loaded data sets
    do.call("rbind", mget(ls(pattern = "^hk_2015_disco_*")))
# reset index
  row.names(hk_2015_dsc_raw) <- NULL
# lower column names
  names(hk_2015_dsc_raw) <- tolower(names(hk_2015_dsc_raw))
  
# clean
  clean_not_geocoded_2015_dsc <-
    hk_2015_dsc_raw %>%
    mutate(
      # add a constituency columns
      constituency = 
        case_when(
          str_detect(district_code, "A") ~ "Central & Western",
          str_detect(district_code, "B") ~ "Wan Chai",
          str_detect(district_code, "C") ~ "Eastern District",
          str_detect(district_code, "D") ~ "Southern District",
          str_detect(district_code, "E") ~ "Yau Tsim Mong District",
          str_detect(district_code, "F") ~ "Sham Shui Po",
          str_detect(district_code, "G") ~ "Kowloon City District",
          str_detect(district_code, "H") ~ "Wong Tai Sin District",
          str_detect(district_code, "J") ~ "Kwun Tong District",
          str_detect(district_code, "K") ~ "Tsuen Wan District",
          str_detect(district_code, "L") ~ "Tuen Mun District",
          str_detect(district_code, "M") ~ "Yuen Long District",
          str_detect(district_code, "N") ~ "North District",
          str_detect(district_code, "P") ~ "Tai Po District",
          str_detect(district_code, "Q") ~ "Sai Kung District",
          str_detect(district_code, "R") ~ "Sha Tin District",
          str_detect(district_code, "S") ~ "Kwai Tsing District",
          str_detect(district_code, "T") ~ "Islands District",
        ) 
    ) %>%
    # add candidate name columns
    left_join(reshaped_hk_2015_disco_candid, by = "district_code")
```

# save a version for geocoding in python

```{r}
write.csv(clean_not_geocoded_2015_dsc, "/Users/jiayili/Dropbox/senior-thesis/01-code/working-data/clean_not_geocoded_2015_dsc.csv")
```

# load the geocoded file 

```{r}
geocoded_2015_disco_raw <-
  read.csv("~/Desktop/honors-thesis-main/00-data/raw-data/raw-election-data/2015_district_council_election/geocoded_disco_2015.csv")
```

# clean the geocoded file

```{r}
clean_geocoded_2015_disco <-
  geocoded_2015_disco_raw %>%
  mutate(
    # adding general info
    year = "2015",
    
    election_type = "District Council Election",
    
    electoral_rule = "First-past-the-post",
    
    # standardizing the constituency variable again
    constituency =
      case_when(
        constituency %in% c("Central & Western", "Wan Chai", "Sham Shui Po") ~ paste(constituency, "District"),
        TRUE ~ as.character(constituency)
      ),
    
    # remove brackets in the `ps_gmaps_coords` column
    ps_gmaps_coords = str_replace_all(ps_gmaps_coords, "\\[|\\]", "")
  ) %>%
  # separate the `gmaps_coords` column into two: lat and lon
  separate(ps_gmaps_coords, into = c("lat", "lon"), sep = ",", remove = FALSE) %>%
  # check for bad coordinates
  # Hong Kong SAR lies between Latitude 22°08' North and 22°35' North, 
  # Longitude 113°49' East and 114°31' East
  mutate(
    bad_coords =
      case_when(
        as.numeric(lat) < 22|as.numeric(lat) > 23|as.numeric(lon) < 113|as.numeric(lon) > 115 ~ "bad",
        TRUE ~ "good"
      )
  ) %>%
  # remove `bad names`
  select(-bad_names) %>%
  # move the columns
  relocate(year, election_type, electoral_rule) %>%
  relocate(ps_gmaps_coords, lat, lon, bad_coords, .after = name_poll_station)
```

# save a version for hand geocoding

```{r}
write.csv(clean_geocoded_2015_disco,
          "~/Desktop/honors-thesis-main/00-data/raw-data/raw-election-data/2015_district_council_election/handcoding_2015_disco.csv",
          row.names = FALSE)
```

# randomly select a sample for auditing (drop `bad_coords` == bad)

```{r}
set.seed(123)
audit_10p_2015_disco <-
    clean_geocoded_2015_disco %>%
    # drop those i manually corrected
    filter(bad_coords != "bad") %>%
    # sample ten percent of polling stations for auditing
    sample_n(49)
```

# adding party allegiance
# note that only the first affiliated organization listed is used

```{r}
# set the path
path <- "/Users/jiayili/Desktop/honors-thesis-main/00-data/raw-data/raw-election-data/2015_district_council_election"

# load hand geocoded data
handgeocoded_2015_disco_raw <-
  read.csv(glue("{path}/handgeocoded_2015_disco.csv"))

# load the party candidate list and political ideology
party_candidates_2015_disco <-
  read.csv(glue("{path}/stacked_party_candidate_info_2015_disco.csv"),
           na.strings = c("", NA)) %>%
  # drop the index column and alias
  select(-X, -Candidate.Name..Alias.) %>%
  # standardize variable names
  select(
    "district_code" = `Constituency.Code`,
    "district_name" = `Constituency`,
    "candidate_number" = `Candidate.No`,
    "party_allegiance" = `Organization.1`,
    "independent" = `INDEPENDENT.CANDIDATE...NON.AFFILIATED.CANDIDATE`
  )

# load political ideologies for parties
ls_poli_ideo <-
  read.csv(glue("{path}/uni_party_list_2015_disco.csv")) %>%
  # standardize the variable for matching
  rename("party_allegiance" = organization_1) %>%
  # make a unique list
  distinct()

# load the vote share list 
candidate_shares_2015_disco <-
  read.csv(glue("{path}/hk2015_district_election_candidate_shares.csv")) %>%
  # keep variables of interest
  select(-X, -No..of.Votes.Received) %>%
  # standardize variable names
  select(
    "district_code" = `Constituency.Code`,
    "district_name" = `Constituency`,
    "candidate_number" = `Candidate.Number`,
    "name_candidate" = `Name.of.Candidate`
  ) %>%
  # standardize
  mutate(
    candidate_number = as.numeric(candidate_number)
  )
```

# join the political ideology data with the party candidate list

```{r}
unmatched_party_candidates <-
  party_candidates_2015_disco %>%
  anti_join(ls_poli_ideo, by = "party_allegiance")

joined_party_candidates <-
  party_candidates_2015_disco %>%
  left_join(ls_poli_ideo, by = "party_allegiance") %>%
  # fix those mistakes caused by an apostrophe
  mutate(
    poli_ideo =
      case_when(
        party_allegiance == "NEW PEOPLE’S PARTY" ~ "pro-democracy",
        party_allegiance == "EAST KOWLOON DISTRICT\nRESIDENTS’ COMMITTEE" ~  "pro-establishment",
        TRUE ~ as.character(poli_ideo)
      ),
    
    # standardize
    district_name = str_to_upper(district_name)
  ) %>%
  # join the share list
  left_join(candidate_shares_2015_disco, by = c("district_code", "district_name", "candidate_number")) %>%
  # reshape the list
  pivot_wider(
    id_cols = c("district_code", "district_name"),
    names_from = "candidate_number",
    values_from = c("party_allegiance", "poli_ideo", "name_candidate")
)
```

# join the election result file

```{r}
clean_handgeocoded_2015_disco_raw <-
  handgeocoded_2015_disco_raw %>%
  left_join(joined_party_candidates, by = c("district_code", "district_name")) %>%
  # separate the `gmaps_coords` column into two: lat and lon
  separate(ps_gmaps_coords, into = c("lat", "lon"), sep = ",", remove = FALSE) %>%
  # reorder the columns
  select(
    year,
    election_type,
    electoral_rule,
    "district" = constituency,
    district_code,
    district_name,
    chinese_name,
    district_name,
    station_code,
    chinese_name_poll_station,
    name_poll_station,
    ps_gmaps_coords,
    lat,
    lon,
    starts_with("votes_candidate_"),
    starts_with("poli_ideo_"),
    starts_with("name_candidate_"),
    starts_with("party_allegiance")
  )
```

# write a final version

```{r}
write.csv(clean_handgeocoded_2015_disco_raw,
          "/Users/jiayili/Desktop/honors-thesis-main/00-data/cleaned-data/election-data/final_clean_2015_disco.csv",
          row.names = FALSE)
```
